combied write_summaries scripts
