+++ MCNP5/.gitignore
@@ -1,2 +1,1 @@
 xsec_data
-summaries
